+++ gltf/src/net/mgsx/gltf/shaders/emissive-only.fs.glsl
@@ -117,8 +117,8 @@ void main() {
 #endif
     
     // final frag color
-#ifdef MANUAL_SRGB
-    out_FragColor.rgb = vec3(pow(color,vec3(1.0/2.2)));
+#ifdef GAMMA_CORRECTION
+    out_FragColor.rgb = vec3(pow(color,vec3(1.0/GAMMA_CORRECTION)));
 #else
     out_FragColor.rgb = color;
 #endif