+++ gltf/src/net/mgsx/gltf/shaders/gdx-pbr.fs.glsl
@@ -388,8 +388,8 @@ void main() {
     vec3 color = baseColor.rgb;
 
     // final frag color
-#ifdef MANUAL_SRGB
-    out_FragColor = vec4(pow(color,vec3(1.0/2.2)), baseColor.a);
+#ifdef GAMMA_CORRECTION
+    out_FragColor = vec4(pow(color,vec3(1.0/GAMMA_CORRECTION)), baseColor.a);
 #else
     out_FragColor = vec4(color, baseColor.a);
 #endif
@@ -619,8 +619,8 @@ void main() {
 
     
     // final frag color
-#ifdef MANUAL_SRGB
-    out_FragColor = vec4(pow(color,vec3(1.0/2.2)), baseColor.a);
+#ifdef GAMMA_CORRECTION
+    out_FragColor = vec4(pow(color,vec3(1.0/GAMMA_CORRECTION)), baseColor.a);
 #else
     out_FragColor = vec4(color, baseColor.a);
 #endif